separated reference form from reference view
